SonyaNazaryshyna/API_app/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   starlette
beanie==1.27.0
    # via -r requirements/requirements.in
build==1.2.2.post1
    # via pip-tools
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   beanie
    #   pip-tools
    #   uvicorn
dnspython==2.7.0
    # via pymongo
fastapi==0.115.3
    # via -r requirements/requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via -r requirements/requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
lazy-model==0.2.0
    # via beanie
motor==3.6.0
    # via beanie
packaging==24.1
    # via build
pip-tools==7.4.1
    # via -r requirements/requirements.in
pydantic==2.9.2
    # via
    #   -r requirements/requirements.in
    #   beanie
    #   fastapi
    #   lazy-model
pydantic-core==2.23.4
    # via pydantic
pymongo==4.9.2
    # via
    #   -r requirements/requirements.in
    #   motor
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r requirements/requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.41.0
    # via
    #   -r requirements/requirements.in
    #   fastapi
toml==0.10.2
    # via beanie
typing-extensions==4.12.2
    # via
    #   beanie
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.32.0
    # via -r requirements/requirements.in
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
